desafio-2/requirements.txt
pandas==2.3.3
psycopg2-binary==2.9.11
python-dotenv==1.1.1
SQLAlchemy==2.0.43
streamlit==1.50.0
plotly==6.3.1
